create independent server app for the express host of the react app
