+++ .github/workflows/merge.yml
@@ -51,6 +51,6 @@ jobs:
           INPUT_DATA_TAGS: latest
           INPUT_API_IMAGES: ${{ steps.login-ecr.outputs.registry }}/dvorac-ukol-api
           INPUT_API_TAGS: latest
-          INPUT_WEB_IMAGES: ${{ steps.login-ecr.outputs.registry }}/dvorac-ukol-web
-          INPUT_WEB_TAGS: latest
-        run: yarn nx run-many --target=container --projects=api,data,web
+          INPUT_SERVER_IMAGES: ${{ steps.login-ecr.outputs.registry }}/dvorac-ukol-web
+          INPUT_SERVER_TAGS: latest
+        run: yarn nx run-many --target=container --projects=api,data,server
