+++ apps/web/.babelrc
@@ -8,7 +8,6 @@
     ]
   ],
   "plugins": [
-    ["styled-components", { "pure": true, "ssr": true }],
-    "react-refresh/babel"
+    ["styled-components", { "pure": true, "ssr": true }]
   ]
 }
