+++ apps/web/project.json
@@ -62,17 +62,6 @@
         }
       }
     },
-    "serve-node": {
-      "executor": "@nx:js/node",
-      "defaultConfiguration": "development",
-      "options": {
-        "buildTarget": "web:build",
-        "port": 8080,
-        "inspect": "inspect",
-        "watch": true
-
-      }
-    },
     "lint": {
       "executor": "@nx/eslint:lint",
       "outputs": ["{options.outputFile}"]
@@ -82,24 +71,6 @@
       "outputs": ["{workspaceRoot}/coverage/apps/web"],
       "options": {
         "jestConfig": "apps/web/jest.config.ts"
-      }
-    },
-    "container": {
-      "executor": "@nx-tools/nx-container:build",
-      "options": {
-        "engine": "docker",
-        "file": "apps/web/Dockerfile",
-        "metadata": {
-          "images": ["ukol/web"],
-          "load": true,
-          "tags": [
-            "type=schedule",
-            "type=ref,event=branch",
-            "type=ref,event=tag",
-            "type=ref,event=pr",
-            "type=sha,prefix=sha-"
-          ]
-        }
       }
     }
   },
