--- apps/web/server/package.json
@@ -1,12 +1,0 @@
-{
-  "name": "server",
-  "version": "1.0.0",
-  "main": "server.js",
-  "scripts": {
-    "start": "node server.js",
-    "debug": "node --inspect=0.0.0.0:9229 server.js"
-  },
-  "dependencies": {
-    "express": "^4.21.2"
-  }
-}
