+++ nx.json
@@ -21,6 +21,11 @@
           "codeCoverage": true
         }
       }
+    },
+    "@nx/webpack:webpack": {
+      "cache": true,
+      "dependsOn": ["^build"],
+      "inputs": ["default", "^default"]
     }
   },
   "generators": {
