+++ package.json
@@ -12,6 +12,7 @@
     "@apollo/client": "^3.6.9",
     "@tanstack/react-query": "^4.2.1",
     "apollo-server-express": "^3.10.0",
+    "axios": "^1.6.0",
     "core-js": "^3.6.5",
     "cors": "^2.8.5",
     "express": "^4.18.1",
@@ -52,6 +53,10 @@
     "@nx/webpack": "19.0.8",
     "@nx/workspace": "19.0.8",
     "@pmmmwh/react-refresh-webpack-plugin": "^0.5.15",
+    "@svgr/webpack": "^8.0.1",
+    "@swc-node/register": "~1.8.0",
+    "@swc/core": "~1.3.85",
+    "@swc/helpers": "~0.5.2",
     "@testing-library/react": "15.0.6",
     "@types/express": "4.17.14",
     "@types/jest": "29.4.4",
@@ -74,6 +79,8 @@
     "eslint-plugin-react-hooks": "4.6.0",
     "jest": "29.4.3",
     "jest-environment-jsdom": "28.1.3",
+    "jest-environment-node": "^29.4.1",
+    "nodemon": "^3.1.9",
     "nx": "19.0.8",
     "prettier": "^2.6.2",
     "react-refresh": "^0.16.0",
